finishes moving the stuff
